fix projectile compiler issue

+++ Assets/Package-v1/Scripts/Combat/Projectile.cs
@@ -56,7 +56,7 @@ namespace Paraverse
             if (other.CompareTag(targetTag))
             {
                 IMobController controller = other.GetComponent<IMobController>();
-                controller.Stats.UpdateCurrentHealth(-damage);
+                controller.Stats.UpdateCurrentHealth((int)-damage);
                 controller.ApplyKnockBack(mob.transform.position);
                 Destroy(gameObject);
             }
